직원등록, 근태관리, 휴가관리 완료  교육이력관리, 자격증관리 진행중.

+++ src/main/java/com/kinder/kindergarten/DTO/Employee/EmployeeDTO.java
@@ -49,6 +49,6 @@ public class EmployeeDTO {
     private LocalDate hireDate;
 
     // 잔여 연차
-    private int annualLeave;
+    private double annualLeave;
 
 }
